+++ app/services/__init__.py
@@ -4,3 +4,4 @@ from .size import size
 from .beverage import beverage
 from .index import index
 from .report import report
+from .base_service import BaseService
