+++ app/services/ingredient.py
@@ -1,7 +1,7 @@
 from app.common.http_methods import GET, POST, PUT
 from flask import Blueprint, jsonify, request
 
-from app.controllers.base_service import BaseService
+from app.services.base_service import BaseService
 
 from ..controllers import IngredientController
